requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
aiohttp==3.9.1
requests==2.31.0
google-generativeai>=0.8.0
python-dotenv
